Simplify directory structure and reduce file count
