--- src/mcp_server_aws_ecs/servers/__init__.py
@@ -1,10 +1,0 @@
-"""
-MCP Servers for AWS ECS operations.
-
-This module provides separate MCP servers for read and write operations.
-"""
-
-from .read_server import create_read_server
-from .write_server import create_write_server
-
-__all__ = ["create_read_server", "create_write_server"]
